correcao da autoexclusao do administrador

+++ src/main/java/com/alvaro/empresas/passagens/resources/FuncionarioResource.java
@@ -1,5 +1,6 @@
 package com.alvaro.empresas.passagens.resources;
 
+import com.alvaro.empresas.passagens.configurations.exceptions.InternalException.GeneralException;
 import com.alvaro.empresas.passagens.dtos.FuncionarioDTO;
 import com.alvaro.empresas.passagens.helpers.beans.UserLoguedComponent;
 import com.alvaro.empresas.passagens.security.dtos.RegisterDtoFuncionario;
@@ -9,6 +10,7 @@ import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
@@ -44,6 +46,8 @@ public class FuncionarioResource {
     @PreAuthorize("hasAnyRole('ROLE_EMPRESA_ADMIN')")
     public ResponseEntity<Object> delete(@PathVariable(value = "idEmpresa") UUID idEmpresa, @PathVariable(value = "email") String email) {
         userLogued.validIfIsMyEmpresa(idEmpresa);
+        if (userLogued.getUserModel().getLogin().equals(email))
+            throw new GeneralException(HttpStatus.CONFLICT, "Usted no puede autodespedirse");
         funcionarioService.delete(email, idEmpresa);
         return ResponseEntity.noContent().build();
     }
